% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comb.binary.R
\name{comb.binary}
\alias{comb.binary}
\title{To get the binary combination of several factors}
\usage{
comb.binary(vector, n)
}
\arguments{
\item{vector}{a vector of factors that you want to be combinated}

\item{n}{the combination number. If n is missing, function comb.binary will return the whole possible combinations}
}
\value{
a matrix
}
\description{
You can use this package to get binary combinations of several factors.Great thanks to Marat Talipov
}
\examples{
comb.binary(1:4)
comb.binary(c(1,3,4),2)
comb.binary(c("a","b","c"))
comb.binary(c("a","b","c"),2)
}
\author{
Jing Zhang
}
