% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oncomine_bar.R
\name{Oncomine_bar}
\alias{Oncomine_bar}
\title{To download bar data
To download bar data frome Oncomine}
\usage{
Oncomine_bar(filename, label_zero = TRUE)
}
\arguments{
\item{filename}{string name}

\item{label_zero}{label zero data}
}
\value{
A dataframe
}
\description{
To download bar data
To download bar data frome Oncomine
}
\examples{
Oncomine_bar("txt.txt")
}
