% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oncomine_bar_plot.R
\name{Oncomine_bar_plot}
\alias{Oncomine_bar_plot}
\title{To plot bar chat}
\usage{
Oncomine_bar_plot(filename)
}
\arguments{
\item{filename}{the filename can be bar data filename or dataframe name of bar or box or omitted}
}
\value{
a bar chart plot by plotly
}
\description{
To plot bar chat
}
\examples{
Oncomine_bar_plot()
Oncomine_bar_plot("bat.txt")
Oncomine_bar_plot(bar_dataframe)
}
