% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oncomine_box.R
\name{Oncomine_box}
\alias{Oncomine_box}
\title{To dowanlaod box data
To dowanlaod Oncomine box data}
\usage{
Oncomine_box(filename, write_csv = FALSE)
}
\arguments{
\item{filename}{string file name}

\item{write_csv}{judge wheter to output box data}
}
\value{
A dataframe
}
\description{
To dowanlaod box data
To dowanlaod Oncomine box data
}
\examples{
Oncomine_box("txt.txt")
}
