% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oncomine_box_plot.R
\name{Oncomine_box_plot}
\alias{Oncomine_box_plot}
\title{To plot box plot for Oncomine}
\usage{
Oncomine_box_plot(filename)
}
\arguments{
\item{filename}{the filename can be box or bar data filename or dataframe name of bar or box or omitted}
}
\value{
a box plot using plotly package
}
\description{
To plot box plot for Oncomine
}
\examples{
filename=Oncomine_box("bb.txt")
filename=Oncomine_bar("bar.txt")
filename="bb.txt"
filename="bar.txt"
Oncomine_box_plot(filename)
}
