% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oncomine_datasets_forGenes.R
\name{Oncomine_datasets_forGenes}
\alias{Oncomine_datasets_forGenes}
\title{To clear Oncomine datasets for Genes}
\usage{
Oncomine_datasets_forGenes(filenames, Simple = TRUE)
}
\arguments{
\item{filenames}{filenames can be one or more or missing}

\item{Simple}{if Simple=TRUE,then to get just data type and study name. If Simple=FALSE, then to get whole information}
}
\value{
a dataframe
}
\description{
To clear Oncomine datasets for Genes
}
\examples{
Oncomine_datasets_forGenes()
}
