% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oncomine_datasets_forGenes_Venn.R
\name{Oncomine_datasets_forGenes_Venn}
\alias{Oncomine_datasets_forGenes_Venn}
\title{To get studys that research the same Genes}
\usage{
Oncomine_datasets_forGenes_Venn(filesname, Plot = TRUE, data = TRUE,
  cexsn = 1, cexil = 1)
}
\arguments{
\item{filesname}{The names of files which contain the data of each Gene. We recomend the file name is identical with Gene.}

\item{Plot}{Plot is a logical object. TRUE to plot Venn and as defaulted. FALSE refuse to do.}

\item{data}{data is a logical object. TRUE, as defaulted, to Get Venn data which includes a binary combination and number of study and study name. FALSE refuse to do.}

\item{cexsn}{to adjust the size of number in venn}

\item{cexil}{to adjust the size of label beside venn}
}
\value{
a dataframe and a venn plot
}
\description{
You can use this function to get the studys that reasearch the sames Genes. This funtion can not be used in Oncomine web for free. We design 2 styles. First is Venn plot, which can be ploted for no more than 7 Genes. Venn data is the other style, Which is suitable for Genes whiout limitation. We also recomend you to view this data in EXCEL.
}
\examples{
Oncomine_datasets_forGenes_Venn(c( "ASXL1.txt", "KRAS.txt",  "SF3B1.txt", "TP53.txt" ))

}
\author{
Jing Zhang
}
