% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oncomine_datasets_forGenes_Volcano.R
\name{Oncomine_datasets_forGenes_Volcano}
\alias{Oncomine_datasets_forGenes_Volcano}
\title{to plot volcano for Genes datasets}
\usage{
Oncomine_datasets_forGenes_Volcano(filenames)
}
\arguments{
\item{filenames}{filenames can be omitted, a list of string}
}
\value{
a plot by plotly
}
\description{
to plot volcano for Genes datasets
}
\examples{
Oncomine_datasets_forGenes_Volcano()
Oncomine_datasets_forGenes_Volcano("1.txt")
Oncomine_datasets_forGenes_Volcano(c("1.txt","2.txt"))
}
