% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oncomine_heatmap_Comparison.of.Concept.R
\name{Oncomine_heatmap_Comparison.of.Concept}
\alias{Oncomine_heatmap_Comparison.of.Concept}
\title{To get data from Comparison of Concept with primary Concept}
\usage{
Oncomine_heatmap_Comparison.of.Concept(filename)
}
\arguments{
\item{filename}{filename can be one or more, or missing if only data files in your working directory.}
}
\value{
a dataframe
}
\description{
To get data from Comparison of Concept with primary Concept
}
\examples{
Oncomine_heatmap_Comparison.of.Concept()
}
