% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Oncomine_heatmap_ComparisonofAllGenesin_SomeDisease.R
\name{Oncomine_heatmap_ComparisonofAllGenesin_SomeDisease}
\alias{Oncomine_heatmap_ComparisonofAllGenesin_SomeDisease}
\title{to clear heat map data of Comparison of All Genes in Some Disease in treatment and control}
\usage{
Oncomine_heatmap_ComparisonofAllGenesin_SomeDisease(filenames)
}
\arguments{
\item{filenames}{filenames can be omitted, a list of string}
}
\value{
a dataframe
}
\description{
to clear heat map data of Comparison of All Genes in Some Disease in treatment and control
}
\examples{
Oncomine_heatmap_ComparisonofAllGenesin_SomeDisease()
Oncomine_heatmap_ComparisonofAllGenesin_SomeDisease("1.txt")
Oncomine_heatmap_ComparisonofAllGenesin_SomeDisease(c("1.txt","2.txt"))
}
