% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Oncomine_heatmap_ComparisonofAllGenesin_SomeDisease_MatrixForHeatmap.R
\name{Oncomine_heatmap_ComparisonofAllGenesin_SomeDisease_MatrixForHeatmap}
\alias{Oncomine_heatmap_ComparisonofAllGenesin_SomeDisease_MatrixForHeatmap}
\title{To prepare a matrix for coexpression data in order to plot heatmap}
\usage{

  Oncomine_heatmap_ComparisonofAllGenesin_SomeDisease_MatrixForHeatmap(filenames,
  rowscale = TRUE)
}
\arguments{
\item{filenames}{can be filenames end with .txt .doc .docx or a dataframe or omitted}

\item{rowscale}{a logical param to scale each row if rowscale=TRUE}
}
\value{
a matrix
}
\description{
you can use function Oncomine_heatmap_ComparisonofAllGenesin_SomeDisease_MatrixForHeatmap() fisrt to get a data frame, and ent to get matrix.
}
\examples{
Oncomine_heatmap_ComparisonofAllGenesin_SomeDisease_MatrixForHeatmap()
Oncomine_heatmap_ComparisonofAllGenesin_SomeDisease_MatrixForHeatmap("1.txt")
Oncomine_heatmap_ComparisonofAllGenesin_SomeDisease_MatrixForHeatmap(c("1.txt","2.txt"))
Oncomine_heatmap_ComparisonofAllGenesin_SomeDisease_MatrixForHeatmap(dataframe)
}
