% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Oncomine_heatmap_ComparisonofAllGenesin_SomeDisease_volcano.R
\name{Oncomine_heatmap_ComparisonofAllGenesin_SomeDisease_volcano}
\alias{Oncomine_heatmap_ComparisonofAllGenesin_SomeDisease_volcano}
\title{to plot volcano heat map data of Comparison of All Genes in Some Disease in treatment and control}
\usage{
Oncomine_heatmap_ComparisonofAllGenesin_SomeDisease_volcano(filenames)
}
\arguments{
\item{filenames}{filenames can be omitted, a list of string}
}
\value{
a plot by plotly
}
\description{
to plot volcano heat map data of Comparison of All Genes in Some Disease in treatment and control
}
\examples{
Oncomine_heatmap_ComparisonofAllGenesin_SomeDisease_volcano()
Oncomine_heatmap_ComparisonofAllGenesin_SomeDisease_volcano("1.txt")
Oncomine_heatmap_ComparisonofAllGenesin_SomeDisease_volcano(c("1.txt","2.txt"))
}
