% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oncomine_heatmap_DiseaseSummaryfor_SomeGene.R
\name{Oncomine_heatmap_DiseaseSummaryfor_SomeGene}
\alias{Oncomine_heatmap_DiseaseSummaryfor_SomeGene}
\title{To dowanload Oncomine heatmap data DiseaseSummaryfor_SomeGene}
\usage{
Oncomine_heatmap_DiseaseSummaryfor_SomeGene(filename)
}
\arguments{
\item{filename}{string filename}
}
\value{
A dataframe
}
\description{
To dowanload Oncomine heatmap data DiseaseSummaryfor_SomeGene
}
\examples{
Oncomine_heatmap_DiseaseSummaryforGene("txt.txt")
}
