% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oncomine_heatmap_Genes_Coexpression.R
\name{Oncomine_heatmap_Genes_Coexpression}
\alias{Oncomine_heatmap_Genes_Coexpression}
\title{To clear Oncomine heatmap Genes Coexpression data including one page and many pages}
\usage{
Oncomine_heatmap_Genes_Coexpression(filenames)
}
\arguments{
\item{filenames}{can be end with .txt .doc .docx or omitted}
}
\value{
a dataframe
}
\description{
To clear Oncomine heatmap Genes Coexpression data including one page and many pages
}
\examples{
data1=Oncomine_heatmap_Genes_Coexpression("3.txt")
data2=Oncomine_heatmap_Genes_Coexpression(c("1.txt","2.txt","3.txt"))
data3=Oncomine_heatmap_Genes_Coexpression()
}
