% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Oncomine_heatmap_Genes_Coexpression_MatrixForHeatmap.R
\name{Oncomine_heatmap_Genes_Coexpression_MatrixForHeatmap}
\alias{Oncomine_heatmap_Genes_Coexpression_MatrixForHeatmap}
\title{To prepare a matrix for coexpression data in order to plot heatmap}
\usage{
Oncomine_heatmap_Genes_Coexpression_MatrixForHeatmap(filenames)
}
\arguments{
\item{filenames}{can be filenames end with .txt .doc .docx or a dataframe or omitted}
}
\value{
a matrix
}
\description{
you can use function Oncomine_heatmap_Genes_Coexpression() fisrt to get a data frame, and ent to get matrix.
}
\examples{
Oncomine_heatmap_Genes_Coexpression_MatrixForHeatmap()
Oncomine_heatmap_Genes_Coexpression_MatrixForHeatmap("1.txt")
Oncomine_heatmap_Genes_Coexpression_MatrixForHeatmap(c("1.txt","2.txt"))
Oncomine_heatmap_Genes_Coexpression_MatrixForHeatmap(dataframe)
}
