% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oncomine_heatmap_Genes_Coexpression_Scatter.R
\name{Oncomine_heatmap_Genes_Coexpression_Scatter}
\alias{Oncomine_heatmap_Genes_Coexpression_Scatter}
\title{To draw a scatter plot for Genes wiht goal Gene using correlation and}
\usage{
Oncomine_heatmap_Genes_Coexpression_Scatter(data_file_name)
}
\arguments{
\item{data_file_name}{This param can be file name end with .txt .doc .docx, or can be omitted}
}
\value{
a scatter plt using plotly
}
\description{
To draw a scatter plot for Genes wiht goal Gene using correlation as x and -log10(p.value) as y.
}
\examples{
Oncomine_heatmap_Genes_Coexpression_Scatter("1.txt")
}
