% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oncomine_meta.R
\name{Oncomine_meta}
\alias{Oncomine_meta}
\title{To download data heatmap ComparssionOfAllGenes for meta analysis
To download data heatmap ComparssionOfAllGenes for meta analysis from oncomine}
\usage{
Oncomine_meta(filename)
}
\arguments{
\item{filename}{string name}
}
\value{
A dataframe
}
\description{
To download data heatmap ComparssionOfAllGenes for meta analysis
To download data heatmap ComparssionOfAllGenes for meta analysis from oncomine
}
\examples{
Oncomine_meta("txt.txt")
}
