% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oncomine_meta_MatrixForHeatmap.R
\name{Oncomine_meta_MatrixForHeatmap}
\alias{Oncomine_meta_MatrixForHeatmap}
\title{To get heat map matrix for meta analysis}
\usage{
Oncomine_meta_MatrixForHeatmap(filename, rowscale = TRUE)
}
\arguments{
\item{filename}{can be .txt .doc .docx or omitted or dataframe acquired by function Oncomine_meta}

\item{rowscale}{a logistic object to scale each row}
}
\value{
an matrix
}
\description{
To get heat map matrix for meta analysis
}
\examples{
Oncomine_meta_MatrixForHeatmap("txt.txt")
}
